refactor: 노드 상세 전체 페이지 시 텝 서버 컴포넌트로 유지할 수 있도록 수정
